refactor(runner): move job lifecycle types
